feat: add Utils. (static func 궁금증)
